Added required DigtialOcean check build scripts (#3)
* added required digtial ocean check scripts

* also cleanup dpkg.log

* pre-check line break and verify

+++ digital-ocean/scripts/90-cleanup.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+# DigitalOcean Marketplace Image Validation Tool
+# © 2021 DigitalOcean LLC.
+# This code is licensed under Apache 2.0 license (see LICENSE.md for details)
+
+set -o errexit
+
 # Ensure /tmp exists and has the proper permissions before
 # checking for security updates
 # https://github.com/digitalocean/marketplace-partners/issues/94
@@ -8,14 +14,21 @@ if [[ ! -d /tmp ]]; then
 fi
 chmod 1777 /tmp
 
-apt-get -y update
-apt-get -y upgrade
+if [ -n "$(command -v yum)" ]; then
+  yum update -y
+  yum clean all
+elif [ -n "$(command -v apt-get)" ]; then
+  export DEBIAN_FRONTEND=noninteractive
+  apt-get -y update
+  apt-get -o Dpkg::Options::="--force-confold" upgrade -q -y --force-yes
+  apt-get -y autoremove
+  apt-get -y autoclean
+fi
+
 rm -rf /tmp/* /var/tmp/*
 history -c
 cat /dev/null > /root/.bash_history
 unset HISTFILE
-apt-get -y autoremove
-apt-get -y autoclean
 find /var/log -mtime -1 -type f -exec truncate -s 0 {} \;
 rm -rf /var/log/*.gz /var/log/*.[0-9] /var/log/*-????????
 rm -rf /var/lib/cloud/instances/*
@@ -30,32 +43,7 @@ printf "\n${GREEN}Writing zeros to the remaining disk space to securely
 erase the unused portion of the file system.
 Depending on your disk size this may take several minutes.
 The secure erase will complete successfully when you see:${NC}
-    writing to '/zerofile': No space left on device\n
-"
+    dd: writing to '/zerofile': No space left on device\n
+Beginning secure erase now\n"
 
-if [[ ! -v IMMICH_TEST_PROD_BRANCH ]]; then 
-    # Time between tests.
-    IMMICH_TEST_PROD_BRANCH="main"
-fi
-
-# Get branch
-IMMICH_BRANCH_REF_NAME=$(cat /opt/immich/branch.txt | tr -d '\n')
-
-if [[ "$IMMICH_BRANCH_REF_NAME" == "" ]]; then 
-    IMMICH_BRANCH_REF_NAME="main"
-fi
-
-echo "on branch $IMMICH_TEST_PROD_BRANCH"
-
-if [ "$IMMICH_BRANCH_REF_NAME" == "$IMMICH_TEST_PROD_BRANCH" ]; then
-  printf "\n${GREEN} Erasing with dd"
-  dd if=/dev/zero of=/zerofile
-else
-  printf "\n${GREEN} Erasing with fallocate"
-  fallocate -l 10G /zerofile
-fi
-
-sync; rm /zerofile; sync
-cat /dev/null > /var/log/lastlog; cat /dev/null > /var/log/wtmp
-sudo apt-get --yes purge droplet-agent*
-
+dd if=/dev/zero of=/zerofile bs=4096 || rm /zerofile
